Refactor map module
